test(golurk): add test for sandstorm damage changes
